+++ packages/keyboard/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @solid-primitives/keyboard
 
+## 1.3.5
+
+### Patch Changes
+
+- Updated dependencies [6680ab9]
+  - @solid-primitives/utils@6.4.0
+  - @solid-primitives/event-listener@2.4.5
+  - @solid-primitives/rootless@1.5.3
+
+## 1.3.4
+
+### Patch Changes
+
+- Updated dependencies [f32f209]
+  - @solid-primitives/event-listener@2.4.4
+
 ## 1.3.3
 
 ### Patch Changes