+++ packages/raf/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @solid-primitives/raf
 
+## 2.3.4
+
+### Patch Changes
+
+- Updated dependencies [6680ab9]
+  - @solid-primitives/utils@6.4.0
+
+## 2.3.3
+
+### Patch Changes
+
+- f32f209: Update author email for David Di Biase.
+
 ## 2.3.2
 
 ### Patch Changes