+++ packages/share/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @solid-primitives/share
 
+## 2.2.4
+
+### Patch Changes
+
+- f32f209: Update author email for David Di Biase.
+
+## 2.2.3
+
+### Patch Changes
+
+- f5abdc3: @solid-primitives/graphql: Fix a build issue occurring during the bundling process.
+  @solid-primitives/share: Fix a typo for Facebook Messenger
+
 ## 2.2.2
 
 ### Patch Changes