+++ packages/trigger/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-primitives/trigger
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies [6680ab9]
+  - @solid-primitives/utils@6.4.0
+
 ## 1.2.2
 
 ### Patch Changes